Streamlining of the API per #446. PR creates a new API extension point in the pipeline ISoapMessageFilter. Change API to be more in line with ASP.NET Core MVC.
ISoapMessageFilter replaces
* IAsyncMessageFilter
* IMessageFilter
* IMessageInspector
* IMessageInspector2

The new API is more future proof, so if there will be a need to add more information, it is possbile to do so by adding a property to the MessageFilterExecutingContext or MessageFilterExecutedContext.

It is also far simple from the conceptual point of view, original API had before and after methods that were called at different points, while the new one .

Shareing state between before and after is now very yeasy thanks to async + next methods that allow filters to keep shared state in the closure of the function.

There are two contexts (executing and executed). It is better to use two different types of contex, so they can evolve independently.